// src/lib/api/__generated/openapi.ts/zod.gen.ts
export const zUpdatePlayerRequest = z.object({
    displayName: z.string().optional(),
    teamIds: z.array(z.string()).optional(),
    timeZone: z.string().optional()
});

export const zCharacterRequest = z.object({
    name: z.string(),
    realm: z.string(),
    region: z.string().default('EU'),
    characterClass: z.string(),
    primarySpec: z.string(),
    eligibleRoles: z.array(z.string()),
    teamIds: z.array(z.string()),
    isMain: z.boolean()
});

export const zVerifyCharacterRequest = z.object({
    approved: z.boolean().optional().default(true),
    note: z.string().nullish()
});
// ...
export const zUpdatePlayerBody = zUpdatePlayerRequest;
// ...
export const zUpdateCharacterBody = zCharacterRequest;

export const zUpdateCharacterPath = z.object({
    characterId: z.string()
});

/**
 * Success
 */
export const zUpdateCharacterResponse = zCharacterEnvelope;

export const zVerifyCharacterBody = zVerifyCharacterRequest;